_random + _locale: extract into module/_random/ + module/_locale/
importing.rs shrinks by 472 lines.
