+++ pyre/pyre-interpreter/src/module/_locale/mod.rs
@@ -1,0 +1,7 @@
+//! _locale module — PyPy: pypy/module/_locale/
+//!
+//! Provides the 'C' locale defaults so locale.py's `from _locale import *`
+//! succeeds and Lib/locale.py exposes working `localeconv` / `setlocale`.
+
+pub mod interp_locale;
+pub mod moduledef;
